Undo/ Redo Improvements and Bug fixing
- Added a new file called annotationEvent that has both Undo and Redo Event.
- Undo and Redo Events derive from pmmEvent so that it can be used by both Spine and Segments
- pmmEventType now have a category property that allows you to distinguish between Spine and Segments easily
- Added unit tests to test undo and redo operations
  -> Some parts of the tests have revealed bugs in the backend that need to be patched. Need to go back and redo the tests when bugs are fixed.

+++ src/pymapmanager/interface2/stackWidgets/stackWidget2.py
@@ -30,14 +30,15 @@ from pymapmanager.timeseriesCore import TimeSeriesCore
 
 from pymapmanager.interface2.stackWidgets.event.spineEvent import (AddSpineEvent, 
                                                                    DeleteSpineEvent,  
-                                                                   UndoSpineEvent,
-                                                                   RedoSpineEvent,
+                                                                #    UndoSpineEvent,
+                                                                #    RedoSpineEvent,
                                                                    EditSpinePropertyEvent)
 
 from pymapmanager.interface2.stackWidgets.event.segmentEvent import (AddSegmentEvent,
                                                                      DeleteSegmentEvent,
                                                                      AddSegmentPoint)
 
+from pymapmanager.interface2.stackWidgets.event.annotationEvent import RedoEvent, UndoEvent
 from pymapmanager._logger import logger
 
 class stackWidget2(mmWidget2):
@@ -271,6 +272,16 @@ class stackWidget2(mmWidget2):
 
     def keyPressEvent(self, event : QtGui.QKeyEvent):
         logger.info(f'{self.getClassName()} {event.text()}')
+
+        # abj: moved undo/ redo shortcuts to stackwidget level
+        if event.modifiers() == QtCore.Qt.ControlModifier and event.key() == QtCore.Qt.Key_Z:
+            logger.info(f"stack widget is pressing z")
+            self.emitUndoEvent()
+
+        if (event.modifiers() == (QtCore.Qt.ControlModifier | QtCore.Qt.ShiftModifier)) and \
+            event.key() == QtCore.Qt.Key_Z:
+            print("Ctrl+Shift+Z pressed")
+            self.emitRedoEvent()
 
         if event.key() == QtCore.Qt.Key_Escape:
             self._cancelSelection()
@@ -668,7 +679,7 @@ class stackWidget2(mmWidget2):
         print('   AFTER addAddSegment')
         print(event)
 
-        # self.getUndoRedo().addUndo(event)
+        self.getUndoRedo().addUndo(event)
 
         return True
 
@@ -680,7 +691,7 @@ class stackWidget2(mmWidget2):
         for segmentID in event.getSegments():
             _deleted = self.getStack().getLineAnnotations().deleteSegment(segmentID)
         
-        # self.getUndoRedo().addUndo(event)
+        self.getUndoRedo().addUndo(event)
         
         return _deleted
     
@@ -704,6 +715,8 @@ class stackWidget2(mmWidget2):
         else:
             self.slot_setStatus('Added point to segment tracing')
         
+        temp = event.getValue("type")
+        logger.info(f" added segment {temp}")
         self.getUndoRedo().addUndo(event)
         
         return _added is not None
@@ -808,9 +821,16 @@ class stackWidget2(mmWidget2):
             # logger.info(f'_rowIdx:{_rowIdx} item:{item}')
             
             deleteSpineID = item['spineID']
-            
+            logger.info(f'deleteSpineID {deleteSpineID}')
+            spineIDExists = self.getStack().getPointAnnotations().spineID_Exists(deleteSpineID)
+
+            # abj: Check if spine exists before deletion
+            if not spineIDExists: 
+                return False
+
             segmentID = [self.getStack().getPointAnnotations().getValue("segmentID", deleteSpineID)]
-            segmentID = segmentID[0]            
+            logger.info(f"segmentID {segmentID}")
+            segmentID = segmentID[0]           
             segmentID = int(segmentID)
 
             _deleted = self.getStack().getPointAnnotations().deleteAnnotation(deleteSpineID)
@@ -1140,33 +1160,42 @@ class stackWidget2(mmWidget2):
         """
         self.move(left,top)
         self.resize(width, height)
-     
-    def undoEvent(self, event : UndoSpineEvent):
+
+    def undoEvent(self, event : UndoEvent):
 
         logger.warning('=== ===   STACK WIDGET PERFORMING Undo   === ===')
 
-        self.getStack().undo()
-        
         undoEvent = self.getUndoRedo().doUndo()
+        # annotationType = undoEvent.category
+        try: # abj
+            annotationType = undoEvent.category
+            logger.info(f"annotationType {annotationType}")
+        
+        except AttributeError:
+            logger.info(f"AttributeError: NoneType annotation")
+            return
+        
+        self.getStack().undo(annotationType)
+        
+        # undoEvent = self.getUndoRedo().doUndo()
         
         event.setUndoEvent(undoEvent)
 
-        # logger.info(f'event:{event}')
-        # logger.info(f'undoEvent:{undoEvent}')
+        logger.info(f'event:{event}')
+        logger.info(f'abj check undoEvent: {undoEvent}')
 
         self.setDirtyTrue() # abj
         
         return undoEvent is not None
     
-    def redoEvent(self, event : RedoSpineEvent):
+    def redoEvent(self, event : RedoEvent):
 
         logger.warning('=== ===   STACK WIDGET PERFORMING Redo   === ===')
 
-        self.getStack().redo()
-
-        # TODO: redo is currently only resetting point annotations 
-        # add support for line annotations
         redoEvent = self.getUndoRedo().doRedo()
+        annotationType = redoEvent.category
+        logger.info(f'redoEvent:{redoEvent}')
+        self.getStack().redo(annotationType)
         
         event.setRedoEvent(redoEvent)
 
@@ -1180,14 +1209,19 @@ class stackWidget2(mmWidget2):
     def emitUndoEvent(self):
         """
         """
-        _undoEvent = UndoSpineEvent(self, None)
-        self.slot_pmmEvent(_undoEvent)
+        # _undoEvent = UndoSpineEvent(self, None)
+        # abj: One undo event for both spines and segments
+        _undoEvent = UndoEvent(self, None)
+        if self.getUndoRedo().numUndo() > 0:
+            self.slot_pmmEvent(_undoEvent)
 
     def emitRedoEvent(self):
         """
         """
-        _redoEvent = RedoSpineEvent(self, None)
-        self.slot_pmmEvent(_redoEvent)
+        # _redoEvent = RedoSpineEvent(self, None)
+        _redoEvent = RedoEvent(self, None)
+        if self.getUndoRedo().numRedo() > 0:
+            self.slot_pmmEvent(_redoEvent)
 
     # abj
     def _old_updateDFwithNewParams(self):
